# backend/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8444
  description: Generated server url
paths:
  /api/inventory/items/{id}/update:
    put:
      tags:
      - items-rest-controller
      summary: Update an item
      operationId: itemUpdating
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: imageField
        in: query
        required: true
        schema:
          type: string
          format: binary
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
        required: true
      responses:
        "404":
          description: No item updated
        "200":
          description: Update an item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
...
  /api/inventory/items:
    post:
      tags:
      - items-rest-controller
      summary: Post a new item
      operationId: addItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
        required: true
      responses:
        "200":
          description: Post a new item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        "404":
          description: No item posted
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Item'
...
  /redis/items/{id}:
    get:
      tags:
      - items-rest-controller
      operationId: getItemByIdReddis
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Item'
...
  /api/inventory/stocks/item/{id}:
    get:
      tags:
      - items-rest-controller
      summary: Get an item stocks
      operationId: getStockItem
      parameters:
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: No item stocks founded
        "200":
          description: Get an item stocks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
...
  /api/inventory/items/{id}:
    get:
      tags:
      - items-rest-controller
      summary: Get an item
      operationId: getItemById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Get an item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        "404":
          description: No item founded
...
  /api/inventory/stocks/{id}/delete:
    delete:
      tags:
      - items-rest-controller
      summary: Delete a stock
      operationId: deleteItemStock
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Delete a stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stock'
        "404":
          description: No stock deleted
  /api/inventory/items/{id}/delete:
    delete:
      tags:
      - items-rest-controller
      summary: Delete an item
      operationId: deleteItem
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Delete an item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        "404":
          description: No item deleted
components:
  schemas:
    Item:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        gender:
          type: string
        type:
          type: string
        imageFile:
          type: object
          properties:
            binaryStream:
              type: object
        stock:
          $ref: '#/components/schemas/StockObject'
        stocks:
          type: array
          items:
            $ref: '#/components/schemas/StockObject'
    Size:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        label:
          type: string
    StockObject:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        stock:
          type: integer
          format: int32
        size:
          $ref: '#/components/schemas/Size'
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    Stock:
      type: object
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        stock:
          type: integer
          format: int32
        size:
          $ref: '#/components/schemas/Size'